+++ shakki/minmaxpaluu.h
@@ -6,6 +6,6 @@
 // Struct ajaisi saman asian. Kun ei rakenneta gettereitä ja settereitä, niin ei tarvita toteutus .cpp tiedostoa
 class MinMaxPaluu{
 public:
-	double _evaluointiArvo;
+	float _evaluointiArvo;
 	Siirto _parasSiirto;
 };
